Issue 2314: Tag set editing
Merging in order to rebase lim's css rewrite against master with the tag sets in. Elz will be reviewing anyway!

+++ lib/tasks/tag_tasks.rake
@@ -27,7 +27,7 @@ namespace :Tag do
   task(:delete_unused => :environment) do
     deleted_names = []
     Tag.find(:all, :conditions => {:canonical => false, :merger_id => nil, :taggings_count => 0}).each do |t|
-      if t.taggings.count == 0 && t.child_taggings.count == 0
+      if t.taggings.count == 0 && t.child_taggings.count == 0 && t.set_taggings.count == 0
         deleted_names << t.name
         t.destroy
       end
